[Bronze V] Title: 팩토리얼 2, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/27433. 팩토리얼 2/팩토리얼 2.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-long long factorial(long long num)
+long long factorial(int num)
 {
     if (num == 1)
         return num;
@@ -12,7 +12,7 @@ long long factorial(long long num)
 
 int main() 
 {
-    long long n;
+    int n;
     cin >> n;
     if (n != 0)
         cout << factorial(n);
